terrain tweaks and :arr:p fuckery
