--- src/main/java/me/hydos/lint/block/organic/LintCorruptGrassBlock.java
@@ -1,58 +1,0 @@
-/*
- * Lint
- * Copyright (C) 2020 hYdos, Valoeghese, ramidzkh
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
-
-package me.hydos.lint.block.organic;
-
-import me.hydos.lint.block.LintBlocks;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.FlowerBlock;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.damage.DamageSource;
-import net.minecraft.entity.effect.StatusEffect;
-import net.minecraft.entity.effect.StatusEffectInstance;
-import net.minecraft.entity.effect.StatusEffects;
-import net.minecraft.entity.mob.HostileEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.BlockView;
-import net.minecraft.world.World;
-
-public class LintCorruptGrassBlock extends FlowerBlock {
-
-	public LintCorruptGrassBlock(StatusEffect effect, Settings settings) {
-		super(effect, 7, settings);
-	}
-
-	@Override
-	public void onEntityCollision(BlockState state, World world, BlockPos pos, Entity entity) {
-		if (entity instanceof LivingEntity && !(entity instanceof PlayerEntity && ((PlayerEntity) entity).isCreative())) {
-			if (!(entity instanceof HostileEntity)) {
-				entity.damage(DamageSource.CACTUS, 1.0F);
-				((LivingEntity) entity).addStatusEffect(new StatusEffectInstance(StatusEffects.BLINDNESS, 30));
-				world.breakBlock(pos, false);
-			}
-		}
-	}
-
-	@Override
-	protected boolean canPlantOnTop(BlockState floor, BlockView view, BlockPos pos) {
-		return floor == LintBlocks.CORRUPT_GRASS.getDefaultState();
-	}
-}
