+++ src/main/java/me/hydos/lint/block/organic/PlaceablePlant.java
@@ -1,0 +1,7 @@
+package me.hydos.lint.block.organic;
+
+import net.minecraft.block.BlockState;
+
+public interface PlaceablePlant {
+	BlockState getGrowsOn();
+}
