+++ src/main/java/me/hydos/lint/world/biome/Biomes.java
@@ -27,7 +27,7 @@ import me.hydos.lint.sound.Sounds;
 import me.hydos.lint.world.biome.surface.DawnShardlandsEdgeSurfaceBuilder;
 import me.hydos.lint.world.biome.surface.DawnShardlandsSurfaceBuilder;
 import me.hydos.lint.world.biome.surface.OceanSurfaceBuilder;
-import me.hydos.lint.world.feature.Features;
+import me.hydos.lint.world.feature.FeaturesOld;
 import me.hydos.lint.world.gen.terrain.TerrainChunkGenerator;
 import me.hydos.lint.world.structure.ConfiguredStructureFeatures;
 import net.fabricmc.fabric.mixin.biome.BuiltinBiomesAccessor;
@@ -131,21 +131,21 @@ public class Biomes {
             .generationSettings(new GenerationSettings.Builder()
                     .surfaceBuilder(MF_SB)
                     .carver(GenerationStep.Carver.AIR, ConfiguredCarvers.CAVE)
-                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, Features.CONFIGURED_FADING_ASH)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.CONFIGURED_RETURN_PORTAL)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.CONFIGURED_TOWN)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.MYSTICAL_TREES)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.MYSTICAL_FLOWERS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.MYSTICAL_GRASS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.GENERIC_BLUE_FLOWERS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.MYSTICAL_STEMS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.TATERBANES)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.SPEARMINTS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.WATERMINTS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.DILLS)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.TARSCAN_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.SICIERON_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.JUREL_ORE)
+                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, FeaturesOld.CONFIGURED_FADING_ASH)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.CONFIGURED_RETURN_PORTAL)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.CONFIGURED_TOWN)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.MYSTICAL_TREES)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.MYSTICAL_FLOWERS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.MYSTICAL_GRASS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.GENERIC_BLUE_FLOWERS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.MYSTICAL_STEMS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.TATERBANES)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.SPEARMINTS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.WATERMINTS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.DILLS)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.TARSCAN_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.SICIERON_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.JUREL_ORE)
                     .feature(GenerationStep.Feature.TOP_LAYER_MODIFICATION, ConfiguredFeatures.FREEZE_TOP_LAYER)
                     .structureFeature(ConfiguredStructureFeatures.DUNGEON)
                     .build())
@@ -169,29 +169,30 @@ public class Biomes {
             .generationSettings(new GenerationSettings.Builder()
                     .surfaceBuilder(MF_SB)
                     .carver(GenerationStep.Carver.AIR, ConfiguredCarvers.CAVE)
-                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, Features.CONFIGURED_FADING_ASH)
-                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, Features.MYSTICAL_ROCKS)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.CONFIGURED_RETURN_PORTAL)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.CONFIGURED_TOWN)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.THICK_MYSTICAL_TREES)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.GENERIC_BLUE_FLOWERS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.MYSTICAL_GRASS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.LESS_MYSTICAL_STEMS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.TATERBANES)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.SPEARMINTS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.WATERMINTS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.TUSSOCKS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.RED_TUSSOCKS)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.TARSCAN_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.SICIERON_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.JUREL_ORE)
+                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, FeaturesOld.CONFIGURED_FADING_ASH)
+                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, FeaturesOld.MYSTICAL_ROCKS)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.CONFIGURED_RETURN_PORTAL)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.CONFIGURED_TOWN)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.THICK_MYSTICAL_TREES)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.GENERIC_BLUE_FLOWERS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.MYSTICAL_GRASS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.LESS_MYSTICAL_STEMS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.TATERBANES)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.SPEARMINTS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.WATERMINTS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.TUSSOCKS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.RED_TUSSOCKS)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.TARSCAN_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.SICIERON_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.JUREL_ORE)
                     .feature(GenerationStep.Feature.TOP_LAYER_MODIFICATION, ConfiguredFeatures.FREEZE_TOP_LAYER)
                     .structureFeature(ConfiguredStructureFeatures.DUNGEON)
                     .build())
             .build();
 
     private static final ConfiguredSurfaceBuilder<TernarySurfaceConfig> FF_SB = SurfaceBuilder.DEFAULT.withConfig(
-            new TernarySurfaceConfig(LintBlocks.FROSTED_GRASS.getDefaultState(),
+            new TernarySurfaceConfig(
+                    LintBlocks.FROSTED_GRASS.getDefaultState(),
                     LintBlocks.RICH_DIRT.getDefaultState(),
                     LintBlocks.RICH_DIRT.getDefaultState()));
 
@@ -213,20 +214,18 @@ public class Biomes {
             .generationSettings(new GenerationSettings.Builder()
                     .surfaceBuilder(FF_SB)
                     .carver(GenerationStep.Carver.AIR, ConfiguredCarvers.CAVE)
-                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, Features.CONFIGURED_FADING_ASH)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.CONFIGURED_RETURN_PORTAL)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.CONFIGURED_TOWN)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.FROZEN_TREES)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.MYSTICAL_FLOWERS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.MYSTICAL_GRASS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.MYSTICAL_STEMS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.TATERBANES)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.SPEARMINTS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.WATERMINTS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.DILLS)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.TARSCAN_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.SICIERON_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.JUREL_ORE)
+                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, FeaturesOld.CONFIGURED_FADING_ASH)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.CONFIGURED_RETURN_PORTAL)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.CONFIGURED_TOWN)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.FROZEN_TREES)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.MARIGOLDS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.FROSTED_TUSSOCKS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.MISTBLOOMS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.FROZEN_STEMS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.SAKHALIN_MINTS)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.TARSCAN_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.SICIERON_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.JUREL_ORE)
                     .feature(GenerationStep.Feature.TOP_LAYER_MODIFICATION, ConfiguredFeatures.FREEZE_TOP_LAYER)
                     .structureFeature(ConfiguredStructureFeatures.DUNGEON)
                     .build())
@@ -258,17 +257,17 @@ public class Biomes {
             .generationSettings(new GenerationSettings.Builder()
                     .surfaceBuilder(CF_SB)
                     .carver(GenerationStep.Carver.AIR, ConfiguredCarvers.CAVE)
-                    .feature(GenerationStep.Feature.RAW_GENERATION, Features.CONFIGURED_VERTICAL_SHAFT)
-                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, Features.CONFIGURED_FADING_ASH)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.CONFIGURED_RETURN_PORTAL)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.CONFIGURED_TOWN)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.CORRUPT_TREES)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.CORRUPT_STEMS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.WILTED_FLOWERS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.KUREI)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.TARSCAN_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.SICIERON_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.JUREL_ORE)
+                    .feature(GenerationStep.Feature.RAW_GENERATION, FeaturesOld.CONFIGURED_VERTICAL_SHAFT)
+                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, FeaturesOld.CONFIGURED_FADING_ASH)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.CONFIGURED_RETURN_PORTAL)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.CONFIGURED_TOWN)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.CORRUPT_TREES)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.CORRUPT_STEMS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.WILTED_FLOWERS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.KUREI)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.TARSCAN_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.SICIERON_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.JUREL_ORE)
                     .feature(GenerationStep.Feature.TOP_LAYER_MODIFICATION, ConfiguredFeatures.FREEZE_TOP_LAYER)
                     .structureFeature(ConfiguredStructureFeatures.DUNGEON)
                     .build()
@@ -298,15 +297,15 @@ public class Biomes {
             .generationSettings(new GenerationSettings.Builder()
                     .surfaceBuilder(OC_SB)
                     .carver(GenerationStep.Carver.AIR, ConfiguredCarvers.CAVE)
-                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, Features.CONFIGURED_FADING_ASH)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.CONFIGURED_TOWN)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.MYSTICAL_STEMS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.MYSTICAL_GRASS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.WATERMINTS)
-                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, Features.DILLS)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.TARSCAN_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.SICIERON_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.JUREL_ORE)
+                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, FeaturesOld.CONFIGURED_FADING_ASH)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.CONFIGURED_TOWN)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.MYSTICAL_STEMS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.MYSTICAL_GRASS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.WATERMINTS)
+                    .feature(GenerationStep.Feature.VEGETAL_DECORATION, FeaturesOld.DILLS)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.TARSCAN_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.SICIERON_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.JUREL_ORE)
                     .feature(GenerationStep.Feature.VEGETAL_DECORATION, ConfiguredFeatures.WARM_OCEAN_VEGETATION)
                     .feature(GenerationStep.Feature.VEGETAL_DECORATION, ConfiguredFeatures.SEAGRASS_WARM)
                     .feature(GenerationStep.Feature.VEGETAL_DECORATION, ConfiguredFeatures.KELP_COLD)
@@ -340,12 +339,12 @@ public class Biomes {
             .generationSettings(new GenerationSettings.Builder()
                     .surfaceBuilder(CB_SB)
                     .carver(GenerationStep.Carver.AIR, ConfiguredCarvers.CAVE)
-                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, Features.CONFIGURED_FADING_ASH)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.CORRUPT_STEMS)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.KUREI)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.TARSCAN_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.SICIERON_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.JUREL_ORE)
+                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, FeaturesOld.CONFIGURED_FADING_ASH)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.CORRUPT_STEMS)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.KUREI)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.TARSCAN_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.SICIERON_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.JUREL_ORE)
                     .feature(GenerationStep.Feature.TOP_LAYER_MODIFICATION, ConfiguredFeatures.FREEZE_TOP_LAYER)
                     .structureFeature(ConfiguredStructureFeatures.DUNGEON)
                     .build())
@@ -375,13 +374,13 @@ public class Biomes {
             .generationSettings(new GenerationSettings.Builder()
                     .surfaceBuilder(IN_SB)
                     .carver(GenerationStep.Carver.AIR, ConfiguredCarvers.CAVE)
-                    .feature(GenerationStep.Feature.RAW_GENERATION, Features.CONFIGURED_VERTICAL_SHAFT)
-                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, Features.CONFIGURED_FADING_ASH)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.CONFIGURED_RETURN_PORTAL)
-                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, Features.CONFIGURED_TOWN)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.TARSCAN_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.SICIERON_ORE)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.JUREL_ORE)
+                    .feature(GenerationStep.Feature.RAW_GENERATION, FeaturesOld.CONFIGURED_VERTICAL_SHAFT)
+                    .feature(GenerationStep.Feature.LOCAL_MODIFICATIONS, FeaturesOld.CONFIGURED_FADING_ASH)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.CONFIGURED_RETURN_PORTAL)
+                    .feature(GenerationStep.Feature.SURFACE_STRUCTURES, FeaturesOld.CONFIGURED_TOWN)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.TARSCAN_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.SICIERON_ORE)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.JUREL_ORE)
                     .feature(GenerationStep.Feature.TOP_LAYER_MODIFICATION, ConfiguredFeatures.FREEZE_TOP_LAYER)
                     .build())
             .build();
@@ -409,7 +408,7 @@ public class Biomes {
             .spawnSettings(new SpawnSettings.Builder().build())
             .generationSettings(new GenerationSettings.Builder()
                     .surfaceBuilder(DS_SB)
-                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, Features.DAWN_SHARDLANDS_SHARDS)
+                    .feature(GenerationStep.Feature.UNDERGROUND_ORES, FeaturesOld.DAWN_SHARDLANDS_SHARDS)
                     // TODO lexmanos boss structure
                     .build())
             .build();
