+++ src/main/java/me/hydos/lint/world/gen/terrain/TerrainChunkGenerator.java
@@ -25,7 +25,7 @@ import me.hydos.lint.Lint;
 import me.hydos.lint.block.LintBlocks;
 import me.hydos.lint.util.callback.ServerChunkManagerCallback;
 import me.hydos.lint.util.math.Vec2i;
-import me.hydos.lint.world.feature.Features;
+import me.hydos.lint.world.feature.FeaturesOld;
 import me.hydos.lint.world.feature.FloatingIslandModifier;
 import me.hydos.lint.world.gen.FraiyaTerrainGenerator;
 import me.hydos.lint.world.structure2.StructureChunkGenerator;
@@ -358,7 +358,7 @@ public class TerrainChunkGenerator extends ChunkGenerator implements StructureCh
 		
 		if (list.size() > n) {
 			List<Supplier<ConfiguredFeature<?, ?>>> toGenerate = new ArrayList<Supplier<ConfiguredFeature<?, ?>>>(list.get(n));
-			toGenerate.add(0, () -> Features.FLOATING_ISLAND_ALLOS_CRYSTAL);
+			toGenerate.add(0, () -> FeaturesOld.FLOATING_ISLAND_ALLOS_CRYSTAL);
 			
 			for (Supplier<ConfiguredFeature<?, ?>> supplier : toGenerate) {
 				try {
