+++ src/main/java/me/hydos/lint/world/tree/CorruptTree.java
@@ -19,7 +19,7 @@
 
 package me.hydos.lint.world.tree;
 
-import me.hydos.lint.world.feature.Features;
+import me.hydos.lint.world.feature.FeaturesOld;
 import net.minecraft.block.sapling.SaplingGenerator;
 import net.minecraft.world.gen.feature.ConfiguredFeature;
 import net.minecraft.world.gen.feature.TreeFeatureConfig;
@@ -29,6 +29,6 @@ import java.util.Random;
 public class CorruptTree extends SaplingGenerator {
     @Override
     protected ConfiguredFeature<TreeFeatureConfig, ?> createTreeFeature(Random random, boolean bl) {
-        return Features.CORRUPT_TREE;
+        return FeaturesOld.CORRUPT_TREE;
     }
 }
