+++ src/main/resources/assets/lint/lang/en_us.json
@@ -80,6 +80,12 @@
   "block.lint.kurei": "Kurei",
   "block.lint.watermint": "Watermint",
   "block.lint.spearmint": "Spearmint",
+  "block.lint.frosted_tussock": "Frosted Tussock",
+  "block.lint.frozen_sapling": "Frozen Sapling",
+  "block.lint.mistbloom": "Mistbloom",
+  "block.lint.marigold": "Marigold",
+  "block.lint.sakhalin_mint": "Sakhalin Mint",
+  "block.lint.frozen_stem": "Frozen Stem",
   "effect.lint.transmutation": "Transmutation",
   "entity.lint.bee_tater": "Bee tater",
   "entity.lint.crab": "Crab",
