+++ src/main/resources/assets/lint/textures/item/suspicious_looking_disc.png
